requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4
numpy==1.16.2
openpyxl==2.6.0
pandas==0.24.1
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
whitenoise==4.1.2
xlrd==1.2.0
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-rest-framework==0.1.0
djangorestframework==3.7.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0